refactor(rag): remove ExtractionResult re-export from extraction/__init__.py
ExtractionResult is defined in app.rag.types and was being re-exported via
extraction/__init__.__all__. Update the 3 importers (chunking.py,
test_chunking.py, test_extraction.py) to import directly from app.rag.types.
The import remains in extraction/__init__.py for its own return-type annotation.
